Change 3 to 5; change precomputation to always on
3 is now used instead of 5 for the generator for powers of 2. This is
the "correct" choice, since it makes certain things much easier.
(Basically, this is because 5 is 1 mod 4, so all powers of 5 are also 1
mod 4.)

This commit also sets precomputation to always happen again. Being able
to do things without precomputation is very important, and we want to
support it, but right now some important things are still broken if we
don't precompute, so it is always turned on.